remove internal generic types (#176)
* remove internal generic types

* refactor: add parquet.Page.Type (#177)

* refactor: add parquet.Page.Type

* refactor indexed page type (#178)

* refactor page index type

* Update dictionary.go


* fix bit-packed encoding of booleans in memory (#179)

* fix bit-packed encoding of booleans in memory

* fix panic on Go 1.17

* fuzz RLE encoding + fix

* Update dictionary.go


* Update encoding/rle/rle.go


* rename: values => bits
