+++ convert.go
@@ -334,7 +334,7 @@ func (p missingPage) Bounds() (min, max Value, ok bool) { return }
 func (p missingPage) Size() int64                       { return 0 }
 func (p missingPage) Values() ValueReader               { return &missingValues{page: p} }
 func (p missingPage) Buffer() BufferedPage {
-	return newErrorPage(p.Column(), "cannot buffer missing page")
+	return newErrorPage(p.Type(), p.Column(), "cannot buffer missing page")
 }
 
 type missingValues struct {
