+++ encoding/rle/testdata/fuzz/FuzzEncodeBoolean/6be5e340694798c2e5b94c758f0262edd2edf8af5795d4c6c60f6e02643bbb96
@@ -1,0 +1,3 @@
+go test fuzz v1
+[]byte("0\x00\x00")
+int64(93)
