+++ page.go
@@ -19,6 +19,13 @@ import (
 //
 // https://github.com/apache/parquet-format#glossary
 type Page interface {
+	// Returns the type of values read from this page.
+	//
+	// The returned type can be used to encode the page data, in the case of
+	// an indexed page (which has a dictionary), the type is configured to
+	// encode the indexes stored in the page rather than the plain values.
+	Type() Type
+
 	// Returns the column index that this page belongs to.
 	Column() int
 
@@ -212,17 +219,20 @@ func forEachPageSlice(page BufferedPage, wantSize int64, do func(BufferedPage) e
 // as a way to ensure that it is not ignored due to being empty when written
 // to a file.
 type errorPage struct {
+	typ         Type
 	err         error
 	columnIndex int
 }
 
-func newErrorPage(columnIndex int, msg string, args ...interface{}) *errorPage {
+func newErrorPage(typ Type, columnIndex int, msg string, args ...interface{}) *errorPage {
 	return &errorPage{
+		typ:         typ,
 		err:         fmt.Errorf(msg, args...),
 		columnIndex: columnIndex,
 	}
 }
 
+func (page *errorPage) Type() Type                        { return page.typ }
 func (page *errorPage) Column() int                       { return page.columnIndex }
 func (page *errorPage) Dictionary() Dictionary            { return nil }
 func (page *errorPage) NumRows() int64                    { return 1 }
@@ -283,6 +293,8 @@ func newOptionalPage(base BufferedPage, maxDefinitionLevel byte, definitionLevel
 		definitionLevels:   definitionLevels,
 	}
 }
+
+func (page *optionalPage) Type() Type { return page.base.Type() }
 
 func (page *optionalPage) Column() int { return page.base.Column() }
 
@@ -345,6 +357,8 @@ func newRepeatedPage(base BufferedPage, maxRepetitionLevel, maxDefinitionLevel b
 		repetitionLevels:   repetitionLevels,
 	}
 }
+
+func (page *repeatedPage) Type() Type { return page.base.Type() }
 
 func (page *repeatedPage) Column() int { return page.base.Column() }
 
@@ -427,19 +441,491 @@ func (page *repeatedPage) Slice(i, j int64) BufferedPage {
 	)
 }
 
+type booleanPage struct {
+	typ         Type
+	bits        []byte
+	offset      int32
+	numValues   int32
+	columnIndex int16
+}
+
+func newBooleanPage(typ Type, columnIndex int16, numValues int32, values []byte) *booleanPage {
+	return &booleanPage{
+		typ:         typ,
+		bits:        values[:bits.ByteCount(uint(numValues))],
+		numValues:   numValues,
+		columnIndex: ^columnIndex,
+	}
+}
+
+func (page *booleanPage) Type() Type { return page.typ }
+
+func (page *booleanPage) Column() int { return int(^page.columnIndex) }
+
+func (page *booleanPage) Dictionary() Dictionary { return nil }
+
+func (page *booleanPage) NumRows() int64 { return int64(page.numValues) }
+
+func (page *booleanPage) NumValues() int64 { return int64(page.numValues) }
+
+func (page *booleanPage) NumNulls() int64 { return 0 }
+
+func (page *booleanPage) Size() int64 { return int64(len(page.bits)) }
+
+func (page *booleanPage) RepetitionLevels() []byte { return nil }
+
+func (page *booleanPage) DefinitionLevels() []byte { return nil }
+
+func (page *booleanPage) Data() []byte { return page.bits }
+
+func (page *booleanPage) Values() ValueReader { return &booleanPageReader{page: page} }
+
+func (page *booleanPage) Buffer() BufferedPage { return page }
+
+func (page *booleanPage) valueAt(i int) bool {
+	j := uint32(int(page.offset)+i) / 8
+	k := uint32(int(page.offset)+i) % 8
+	return ((page.bits[j] >> k) & 1) != 0
+}
+
+func (page *booleanPage) min() bool {
+	for i := 0; i < int(page.numValues); i++ {
+		if !page.valueAt(i) {
+			return false
+		}
+	}
+	return page.numValues > 0
+}
+
+func (page *booleanPage) max() bool {
+	for i := 0; i < int(page.numValues); i++ {
+		if page.valueAt(i) {
+			return true
+		}
+	}
+	return false
+}
+
+func (page *booleanPage) bounds() (min, max bool) {
+	hasFalse, hasTrue := false, false
+
+	for i := 0; i < int(page.numValues); i++ {
+		v := page.valueAt(i)
+		if v {
+			hasTrue = true
+		} else {
+			hasFalse = true
+		}
+		if hasTrue && hasFalse {
+			break
+		}
+	}
+
+	min = !hasFalse
+	max = hasTrue
+	return min, max
+}
+
+func (page *booleanPage) Bounds() (min, max Value, ok bool) {
+	if ok = page.numValues > 0; ok {
+		minBool, maxBool := page.bounds()
+		min = makeValueBoolean(minBool)
+		max = makeValueBoolean(maxBool)
+	}
+	return min, max, ok
+}
+
+func (page *booleanPage) Clone() BufferedPage {
+	return &booleanPage{
+		typ:         page.typ,
+		bits:        append([]byte{}, page.bits...),
+		offset:      page.offset,
+		numValues:   page.numValues,
+		columnIndex: page.columnIndex,
+	}
+}
+
+func (page *booleanPage) Slice(i, j int64) BufferedPage {
+	off := i / 8
+	end := j / 8
+
+	if (j % 8) != 0 {
+		end++
+	}
+
+	return &booleanPage{
+		typ:         page.typ,
+		bits:        page.bits[off:end],
+		offset:      int32(i % 8),
+		numValues:   int32(j - i),
+		columnIndex: page.columnIndex,
+	}
+}
+
+type int32Page struct {
+	typ         Type
+	values      []int32
+	columnIndex int16
+}
+
+func newInt32Page(typ Type, columnIndex int16, numValues int32, values []byte) *int32Page {
+	return &int32Page{
+		typ:         typ,
+		values:      bits.BytesToInt32(values)[:numValues],
+		columnIndex: ^columnIndex,
+	}
+}
+
+func (page *int32Page) Type() Type { return page.typ }
+
+func (page *int32Page) Column() int { return int(^page.columnIndex) }
+
+func (page *int32Page) Dictionary() Dictionary { return nil }
+
+func (page *int32Page) NumRows() int64 { return int64(len(page.values)) }
+
+func (page *int32Page) NumValues() int64 { return int64(len(page.values)) }
+
+func (page *int32Page) NumNulls() int64 { return 0 }
+
+func (page *int32Page) Size() int64 { return 4 * int64(len(page.values)) }
+
+func (page *int32Page) RepetitionLevels() []byte { return nil }
+
+func (page *int32Page) DefinitionLevels() []byte { return nil }
+
+func (page *int32Page) Data() []byte { return bits.Int32ToBytes(page.values) }
+
+func (page *int32Page) Values() ValueReader { return &int32PageReader{page: page} }
+
+func (page *int32Page) Buffer() BufferedPage { return page }
+
+func (page *int32Page) min() int32 { return bits.MinInt32(page.values) }
+
+func (page *int32Page) max() int32 { return bits.MaxInt32(page.values) }
+
+func (page *int32Page) bounds() (min, max int32) { return bits.MinMaxInt32(page.values) }
+
+func (page *int32Page) Bounds() (min, max Value, ok bool) {
+	if ok = len(page.values) > 0; ok {
+		minInt32, maxInt32 := page.bounds()
+		min = makeValueInt32(minInt32)
+		max = makeValueInt32(maxInt32)
+	}
+	return min, max, ok
+}
+
+func (page *int32Page) Clone() BufferedPage {
+	return &int32Page{
+		typ:         page.typ,
+		values:      append([]int32{}, page.values...),
+		columnIndex: page.columnIndex,
+	}
+}
+
+func (page *int32Page) Slice(i, j int64) BufferedPage {
+	return &int32Page{
+		typ:         page.typ,
+		values:      page.values[i:j],
+		columnIndex: page.columnIndex,
+	}
+}
+
+type int64Page struct {
+	typ         Type
+	values      []int64
+	columnIndex int16
+}
+
+func newInt64Page(typ Type, columnIndex int16, numValues int32, values []byte) *int64Page {
+	return &int64Page{
+		typ:         typ,
+		values:      bits.BytesToInt64(values)[:numValues],
+		columnIndex: ^columnIndex,
+	}
+}
+
+func (page *int64Page) Type() Type { return page.typ }
+
+func (page *int64Page) Column() int { return int(^page.columnIndex) }
+
+func (page *int64Page) Dictionary() Dictionary { return nil }
+
+func (page *int64Page) NumRows() int64 { return int64(len(page.values)) }
+
+func (page *int64Page) NumValues() int64 { return int64(len(page.values)) }
+
+func (page *int64Page) NumNulls() int64 { return 0 }
+
+func (page *int64Page) Size() int64 { return 8 * int64(len(page.values)) }
+
+func (page *int64Page) RepetitionLevels() []byte { return nil }
+
+func (page *int64Page) DefinitionLevels() []byte { return nil }
+
+func (page *int64Page) Data() []byte { return bits.Int64ToBytes(page.values) }
+
+func (page *int64Page) Values() ValueReader { return &int64PageReader{page: page} }
+
+func (page *int64Page) Buffer() BufferedPage { return page }
+
+func (page *int64Page) min() int64 { return bits.MinInt64(page.values) }
+
+func (page *int64Page) max() int64 { return bits.MaxInt64(page.values) }
+
+func (page *int64Page) bounds() (min, max int64) { return bits.MinMaxInt64(page.values) }
+
+func (page *int64Page) Bounds() (min, max Value, ok bool) {
+	if ok = len(page.values) > 0; ok {
+		minInt64, maxInt64 := page.bounds()
+		min = makeValueInt64(minInt64)
+		max = makeValueInt64(maxInt64)
+	}
+	return min, max, ok
+}
+
+func (page *int64Page) Clone() BufferedPage {
+	return &int64Page{
+		typ:         page.typ,
+		values:      append([]int64{}, page.values...),
+		columnIndex: page.columnIndex,
+	}
+}
+
+func (page *int64Page) Slice(i, j int64) BufferedPage {
+	return &int64Page{
+		typ:         page.typ,
+		values:      page.values[i:j],
+		columnIndex: page.columnIndex,
+	}
+}
+
+type int96Page struct {
+	typ         Type
+	values      []deprecated.Int96
+	columnIndex int16
+}
+
+func newInt96Page(typ Type, columnIndex int16, numValues int32, values []byte) *int96Page {
+	return &int96Page{
+		typ:         typ,
+		values:      deprecated.BytesToInt96(values)[:numValues],
+		columnIndex: ^columnIndex,
+	}
+}
+
+func (page *int96Page) Type() Type { return page.typ }
+
+func (page *int96Page) Column() int { return int(^page.columnIndex) }
+
+func (page *int96Page) Dictionary() Dictionary { return nil }
+
+func (page *int96Page) NumRows() int64 { return int64(len(page.values)) }
+
+func (page *int96Page) NumValues() int64 { return int64(len(page.values)) }
+
+func (page *int96Page) NumNulls() int64 { return 0 }
+
+func (page *int96Page) Size() int64 { return 12 * int64(len(page.values)) }
+
+func (page *int96Page) RepetitionLevels() []byte { return nil }
+
+func (page *int96Page) DefinitionLevels() []byte { return nil }
+
+func (page *int96Page) Data() []byte { return deprecated.Int96ToBytes(page.values) }
+
+func (page *int96Page) Values() ValueReader { return &int96PageReader{page: page} }
+
+func (page *int96Page) Buffer() BufferedPage { return page }
+
+func (page *int96Page) min() deprecated.Int96 { return deprecated.MinInt96(page.values) }
+
+func (page *int96Page) max() deprecated.Int96 { return deprecated.MaxInt96(page.values) }
+
+func (page *int96Page) bounds() (min, max deprecated.Int96) {
+	return deprecated.MinMaxInt96(page.values)
+}
+
+func (page *int96Page) Bounds() (min, max Value, ok bool) {
+	if ok = len(page.values) > 0; ok {
+		minInt96, maxInt96 := page.bounds()
+		min = makeValueInt96(minInt96)
+		max = makeValueInt96(maxInt96)
+	}
+	return min, max, ok
+}
+
+func (page *int96Page) Clone() BufferedPage {
+	return &int96Page{
+		typ:         page.typ,
+		values:      append([]deprecated.Int96{}, page.values...),
+		columnIndex: page.columnIndex,
+	}
+}
+
+func (page *int96Page) Slice(i, j int64) BufferedPage {
+	return &int96Page{
+		typ:         page.typ,
+		values:      page.values[i:j],
+		columnIndex: page.columnIndex,
+	}
+}
+
+type floatPage struct {
+	typ         Type
+	values      []float32
+	columnIndex int16
+}
+
+func newFloatPage(typ Type, columnIndex int16, numValues int32, values []byte) *floatPage {
+	return &floatPage{
+		typ:         typ,
+		values:      bits.BytesToFloat32(values)[:numValues],
+		columnIndex: ^columnIndex,
+	}
+}
+
+func (page *floatPage) Type() Type { return page.typ }
+
+func (page *floatPage) Column() int { return int(^page.columnIndex) }
+
+func (page *floatPage) Dictionary() Dictionary { return nil }
+
+func (page *floatPage) NumRows() int64 { return int64(len(page.values)) }
+
+func (page *floatPage) NumValues() int64 { return int64(len(page.values)) }
+
+func (page *floatPage) NumNulls() int64 { return 0 }
+
+func (page *floatPage) Size() int64 { return 4 * int64(len(page.values)) }
+
+func (page *floatPage) RepetitionLevels() []byte { return nil }
+
+func (page *floatPage) DefinitionLevels() []byte { return nil }
+
+func (page *floatPage) Data() []byte { return bits.Float32ToBytes(page.values) }
+
+func (page *floatPage) Values() ValueReader { return &floatPageReader{page: page} }
+
+func (page *floatPage) Buffer() BufferedPage { return page }
+
+func (page *floatPage) min() float32 { return bits.MinFloat32(page.values) }
+
+func (page *floatPage) max() float32 { return bits.MaxFloat32(page.values) }
+
+func (page *floatPage) bounds() (min, max float32) { return bits.MinMaxFloat32(page.values) }
+
+func (page *floatPage) Bounds() (min, max Value, ok bool) {
+	if ok = len(page.values) > 0; ok {
+		minFloat32, maxFloat32 := page.bounds()
+		min = makeValueFloat(minFloat32)
+		max = makeValueFloat(maxFloat32)
+	}
+	return min, max, ok
+}
+
+func (page *floatPage) Clone() BufferedPage {
+	return &floatPage{
+		typ:         page.typ,
+		values:      append([]float32{}, page.values...),
+		columnIndex: page.columnIndex,
+	}
+}
+
+func (page *floatPage) Slice(i, j int64) BufferedPage {
+	return &floatPage{
+		typ:         page.typ,
+		values:      page.values[i:j],
+		columnIndex: page.columnIndex,
+	}
+}
+
+type doublePage struct {
+	typ         Type
+	values      []float64
+	columnIndex int16
+}
+
+func newDoublePage(typ Type, columnIndex int16, numValues int32, values []byte) *doublePage {
+	return &doublePage{
+		typ:         typ,
+		values:      bits.BytesToFloat64(values)[:numValues],
+		columnIndex: ^columnIndex,
+	}
+}
+
+func (page *doublePage) Type() Type { return page.typ }
+
+func (page *doublePage) Column() int { return int(^page.columnIndex) }
+
+func (page *doublePage) Dictionary() Dictionary { return nil }
+
+func (page *doublePage) NumRows() int64 { return int64(len(page.values)) }
+
+func (page *doublePage) NumValues() int64 { return int64(len(page.values)) }
+
+func (page *doublePage) NumNulls() int64 { return 0 }
+
+func (page *doublePage) Size() int64 { return 8 * int64(len(page.values)) }
+
+func (page *doublePage) RepetitionLevels() []byte { return nil }
+
+func (page *doublePage) DefinitionLevels() []byte { return nil }
+
+func (page *doublePage) Data() []byte { return bits.Float64ToBytes(page.values) }
+
+func (page *doublePage) Values() ValueReader { return &doublePageReader{page: page} }
+
+func (page *doublePage) Buffer() BufferedPage { return page }
+
+func (page *doublePage) min() float64 { return bits.MinFloat64(page.values) }
+
+func (page *doublePage) max() float64 { return bits.MaxFloat64(page.values) }
+
+func (page *doublePage) bounds() (min, max float64) { return bits.MinMaxFloat64(page.values) }
+
+func (page *doublePage) Bounds() (min, max Value, ok bool) {
+	if ok = len(page.values) > 0; ok {
+		minFloat64, maxFloat64 := page.bounds()
+		min = makeValueDouble(minFloat64)
+		max = makeValueDouble(maxFloat64)
+	}
+	return min, max, ok
+}
+
+func (page *doublePage) Clone() BufferedPage {
+	return &doublePage{
+		typ:         page.typ,
+		values:      append([]float64{}, page.values...),
+		columnIndex: page.columnIndex,
+	}
+}
+
+func (page *doublePage) Slice(i, j int64) BufferedPage {
+	return &doublePage{
+		typ:         page.typ,
+		values:      page.values[i:j],
+		columnIndex: page.columnIndex,
+	}
+}
+
 type byteArrayPage struct {
+	typ         Type
 	values      []byte
 	numValues   int32
 	columnIndex int16
 }
 
-func newByteArrayPage(columnIndex int16, numValues int32, values []byte) *byteArrayPage {
+func newByteArrayPage(typ Type, columnIndex int16, numValues int32, values []byte) *byteArrayPage {
 	return &byteArrayPage{
+		typ:         typ,
 		values:      values,
 		numValues:   numValues,
 		columnIndex: ^columnIndex,
 	}
 }
+
+func (page *byteArrayPage) Type() Type { return page.typ }
 
 func (page *byteArrayPage) Column() int { return int(^page.columnIndex) }
 
@@ -545,6 +1031,7 @@ func (page *byteArrayPage) cloneValues() []byte {
 
 func (page *byteArrayPage) Clone() BufferedPage {
 	return &byteArrayPage{
+		typ:         page.typ,
 		values:      page.cloneValues(),
 		numValues:   page.numValues,
 		columnIndex: page.columnIndex,
@@ -570,6 +1057,7 @@ func (page *byteArrayPage) Slice(i, j int64) BufferedPage {
 	}
 
 	return &byteArrayPage{
+		typ:         page.typ,
 		values:      page.values[off0:off1:off1],
 		numValues:   int32(numValues),
 		columnIndex: page.columnIndex,
@@ -577,12 +1065,14 @@ func (page *byteArrayPage) Slice(i, j int64) BufferedPage {
 }
 
 type fixedLenByteArrayPage struct {
+	typ         Type
 	data        []byte
 	size        int
 	columnIndex int16
 }
 
-func newFixedLenByteArrayPage(columnIndex int16, numValues int32, data []byte, size int) *fixedLenByteArrayPage {
+func newFixedLenByteArrayPage(typ Type, columnIndex int16, numValues int32, data []byte) *fixedLenByteArrayPage {
+	size := typ.Length()
 	if (len(data) % size) != 0 {
 		panic("cannot create fixed-length byte array page from input which is not a multiple of the type size")
 	}
@@ -590,11 +1080,14 @@ func newFixedLenByteArrayPage(columnIndex int16, numValues int32, data []byte, s
 		panic(fmt.Errorf("number of values mismatch in numValues and data arguments: %d != %d", numValues, len(data)/size))
 	}
 	return &fixedLenByteArrayPage{
+		typ:         typ,
 		data:        data,
 		size:        size,
 		columnIndex: ^columnIndex,
 	}
 }
+
+func (page *fixedLenByteArrayPage) Type() Type { return page.typ }
 
 func (page *fixedLenByteArrayPage) Column() int { return int(^page.columnIndex) }
 
@@ -643,6 +1136,7 @@ func (page *fixedLenByteArrayPage) Bounds() (min, max Value, ok bool) {
 
 func (page *fixedLenByteArrayPage) Clone() BufferedPage {
 	return &fixedLenByteArrayPage{
+		typ:         page.typ,
 		data:        append([]byte{}, page.data...),
 		size:        page.size,
 		columnIndex: page.columnIndex,
@@ -651,39 +1145,181 @@ func (page *fixedLenByteArrayPage) Clone() BufferedPage {
 
 func (page *fixedLenByteArrayPage) Slice(i, j int64) BufferedPage {
 	return &fixedLenByteArrayPage{
+		typ:         page.typ,
 		data:        page.data[i*int64(page.size) : j*int64(page.size)],
 		size:        page.size,
 		columnIndex: page.columnIndex,
 	}
 }
 
+type uint32Page struct {
+	typ         Type
+	values      []uint32
+	columnIndex int16
+}
+
+func newUint32Page(typ Type, columnIndex int16, numValues int32, values []byte) *uint32Page {
+	return &uint32Page{
+		typ:         typ,
+		values:      bits.BytesToUint32(values)[:numValues],
+		columnIndex: ^columnIndex,
+	}
+}
+
+func (page *uint32Page) Type() Type { return page.typ }
+
+func (page *uint32Page) Column() int { return int(^page.columnIndex) }
+
+func (page *uint32Page) Dictionary() Dictionary { return nil }
+
+func (page *uint32Page) NumRows() int64 { return int64(len(page.values)) }
+
+func (page *uint32Page) NumValues() int64 { return int64(len(page.values)) }
+
+func (page *uint32Page) NumNulls() int64 { return 0 }
+
+func (page *uint32Page) Size() int64 { return 4 * int64(len(page.values)) }
+
+func (page *uint32Page) RepetitionLevels() []byte { return nil }
+
+func (page *uint32Page) DefinitionLevels() []byte { return nil }
+
+func (page *uint32Page) Data() []byte { return bits.Uint32ToBytes(page.values) }
+
+func (page *uint32Page) Values() ValueReader { return &uint32PageReader{page: page} }
+
+func (page *uint32Page) Buffer() BufferedPage { return page }
+
+func (page *uint32Page) min() uint32 { return bits.MinUint32(page.values) }
+
+func (page *uint32Page) max() uint32 { return bits.MaxUint32(page.values) }
+
+func (page *uint32Page) bounds() (min, max uint32) { return bits.MinMaxUint32(page.values) }
+
+func (page *uint32Page) Bounds() (min, max Value, ok bool) {
+	if ok = len(page.values) > 0; ok {
+		minUint32, maxUint32 := page.bounds()
+		min = makeValueUint32(minUint32)
+		max = makeValueUint32(maxUint32)
+	}
+	return min, max, ok
+}
+
+func (page *uint32Page) Clone() BufferedPage {
+	return &uint32Page{
+		typ:         page.typ,
+		values:      append([]uint32{}, page.values...),
+		columnIndex: page.columnIndex,
+	}
+}
+
+func (page *uint32Page) Slice(i, j int64) BufferedPage {
+	return &uint32Page{
+		typ:         page.typ,
+		values:      page.values[i:j],
+		columnIndex: page.columnIndex,
+	}
+}
+
+type uint64Page struct {
+	typ         Type
+	values      []uint64
+	columnIndex int16
+}
+
+func newUint64Page(typ Type, columnIndex int16, numValues int32, values []byte) *uint64Page {
+	return &uint64Page{
+		typ:         typ,
+		values:      bits.BytesToUint64(values)[:numValues],
+		columnIndex: ^columnIndex,
+	}
+}
+
+func (page *uint64Page) Type() Type { return page.typ }
+
+func (page *uint64Page) Column() int { return int(^page.columnIndex) }
+
+func (page *uint64Page) Dictionary() Dictionary { return nil }
+
+func (page *uint64Page) NumRows() int64 { return int64(len(page.values)) }
+
+func (page *uint64Page) NumValues() int64 { return int64(len(page.values)) }
+
+func (page *uint64Page) NumNulls() int64 { return 0 }
+
+func (page *uint64Page) Size() int64 { return 8 * int64(len(page.values)) }
+
+func (page *uint64Page) RepetitionLevels() []byte { return nil }
+
+func (page *uint64Page) DefinitionLevels() []byte { return nil }
+
+func (page *uint64Page) Data() []byte { return bits.Uint64ToBytes(page.values) }
+
+func (page *uint64Page) Values() ValueReader { return &uint64PageReader{page: page} }
+
+func (page *uint64Page) Buffer() BufferedPage { return page }
+
+func (page *uint64Page) min() uint64 { return bits.MinUint64(page.values) }
+
+func (page *uint64Page) max() uint64 { return bits.MaxUint64(page.values) }
+
+func (page *uint64Page) bounds() (min, max uint64) { return bits.MinMaxUint64(page.values) }
+
+func (page *uint64Page) Bounds() (min, max Value, ok bool) {
+	if ok = len(page.values) > 0; ok {
+		minUint64, maxUint64 := page.bounds()
+		min = makeValueUint64(minUint64)
+		max = makeValueUint64(maxUint64)
+	}
+	return min, max, ok
+}
+
+func (page *uint64Page) Clone() BufferedPage {
+	return &uint64Page{
+		typ:         page.typ,
+		values:      append([]uint64{}, page.values...),
+		columnIndex: page.columnIndex,
+	}
+}
+
+func (page *uint64Page) Slice(i, j int64) BufferedPage {
+	return &uint64Page{
+		typ:         page.typ,
+		values:      page.values[i:j],
+		columnIndex: page.columnIndex,
+	}
+}
+
 type nullPage struct {
+	typ    Type
 	column int
 	count  int
 }
 
-func newNullPage(columnIndex int16, numValues int32) *nullPage {
+func newNullPage(typ Type, columnIndex int16, numValues int32) *nullPage {
 	return &nullPage{
+		typ:    typ,
 		column: int(columnIndex),
 		count:  int(numValues),
 	}
 }
 
-func (p *nullPage) Column() int                       { return p.column }
-func (p *nullPage) Dictionary() Dictionary            { return nil }
-func (p *nullPage) NumRows() int64                    { return int64(p.count) }
-func (p *nullPage) NumValues() int64                  { return int64(p.count) }
-func (p *nullPage) NumNulls() int64                   { return int64(p.count) }
-func (p *nullPage) Bounds() (min, max Value, ok bool) { return }
-func (p *nullPage) Size() int64                       { return 1 }
-func (p *nullPage) Values() ValueReader {
-	return &nullPageReader{column: p.column, remain: p.count}
+func (page *nullPage) Type() Type                        { return page.typ }
+func (page *nullPage) Column() int                       { return page.column }
+func (page *nullPage) Dictionary() Dictionary            { return nil }
+func (page *nullPage) NumRows() int64                    { return int64(page.count) }
+func (page *nullPage) NumValues() int64                  { return int64(page.count) }
+func (page *nullPage) NumNulls() int64                   { return int64(page.count) }
+func (page *nullPage) Bounds() (min, max Value, ok bool) { return }
+func (page *nullPage) Size() int64                       { return 1 }
+func (page *nullPage) Values() ValueReader {
+	return &nullPageReader{column: page.column, remain: page.count}
 }
-func (p *nullPage) Buffer() BufferedPage { return p }
-func (p *nullPage) Clone() BufferedPage  { return p }
-func (p *nullPage) Slice(i, j int64) BufferedPage {
-	return &nullPage{column: p.column, count: p.count - int(j-i)}
+func (page *nullPage) Buffer() BufferedPage { return page }
+func (page *nullPage) Clone() BufferedPage  { return page }
+func (page *nullPage) Slice(i, j int64) BufferedPage {
+	return &nullPage{column: page.column, count: page.count - int(j-i)}
 }
-func (p *nullPage) RepetitionLevels() []byte { return nil }
-func (p *nullPage) DefinitionLevels() []byte { return nil }
-func (p *nullPage) Data() []byte             { return nil }
+func (page *nullPage) RepetitionLevels() []byte { return nil }
+func (page *nullPage) DefinitionLevels() []byte { return nil }
+func (page *nullPage) Data() []byte             { return nil }
