+++ writer.go
@@ -302,7 +302,6 @@ func newWriter(output io.Writer, config *WriterConfig) *writer {
 		c := &writerColumn{
 			buffers:            buffers,
 			pool:               config.ColumnPageBuffers,
-			valueType:          leaf.node.Type(),
 			columnPath:         leaf.path,
 			columnType:         columnType,
 			columnIndex:        columnType.NewColumnIndexer(config.ColumnIndexSizeLimit),
@@ -733,8 +732,10 @@ func (wb *writerBuffers) prependLevelsToDataPageV1(maxRepetitionLevel, maxDefini
 	}
 }
 
-func (wb *writerBuffers) encode(typ Type, page BufferedPage, enc encoding.Encoding) (err error) {
-	wb.page, err = typ.Encode(wb.page[:0], page.Data(), enc)
+func (wb *writerBuffers) encode(page BufferedPage, enc encoding.Encoding) (err error) {
+	pageType := page.Type()
+	pageData := page.Data()
+	wb.page, err = pageType.Encode(wb.page[:0], pageData, enc)
 	return err
 }
 
@@ -756,9 +757,8 @@ type writerColumn struct {
 	pool  PageBufferPool
 	pages []io.ReadWriter
 
-	valueType    Type
-	columnType   Type
 	columnPath   columnPath
+	columnType   Type
 	columnIndex  ColumnIndexer
 	columnBuffer ColumnBuffer
 	columnFilter BloomFilterColumn
@@ -1047,12 +1047,7 @@ func (c *writerColumn) writeBufferedPage(page BufferedPage) (int64, error) {
 		buf.encodeDefinitionLevels(page, c.maxDefinitionLevel)
 	}
 
-	hasDict := page.Dictionary() != nil
-	pageType := c.columnType
-	if !hasDict {
-		pageType = c.valueType
-	}
-	if err := buf.encode(pageType, page, c.page.encoding); err != nil {
+	if err := buf.encode(page, c.page.encoding); err != nil {
 		return 0, fmt.Errorf("encoding parquet data page: %w", err)
 	}
 	if c.dataPageType == format.DataPage {
@@ -1066,7 +1061,7 @@ func (c *writerColumn) writeBufferedPage(page BufferedPage) (int64, error) {
 		}
 	}
 
-	if !hasDict {
+	if page.Dictionary() == nil {
 		switch {
 		case len(c.filter.bits) > 0:
 			// When the writer knows the number of values in advance (e.g. when
@@ -1222,7 +1217,7 @@ func (c *writerColumn) writeDictionaryPage(output io.Writer, dict Dictionary) (e
 	buf := c.buffers
 	buf.reset()
 
-	if err := buf.encode(c.valueType, dict.Page(), &Plain); err != nil {
+	if err := buf.encode(dict.Page(), &Plain); err != nil {
 		return fmt.Errorf("writing parquet dictionary page: %w", err)
 	}
 
@@ -1261,7 +1256,9 @@ func (c *writerColumn) writeDictionaryPage(output io.Writer, dict Dictionary) (e
 }
 
 func (w *writerColumn) writePageToFilter(page BufferedPage) (err error) {
-	w.filter.bits, err = w.valueType.Encode(w.filter.bits, page.Data(), w.columnFilter.Encoding())
+	pageType := page.Type()
+	pageData := page.Data()
+	w.filter.bits, err = pageType.Encode(w.filter.bits, pageData, w.columnFilter.Encoding())
 	return err
 }
